backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
pydantic[email]
jinja2
aiofiles
python-multipart
alembic
passlib[bcrypt]
python-dotenv
PyJWT
psycopg2-binary


# For PDF receipts
pdfkit
reportlab
qrcode
pillow
# wkhtmltopdf-binary

# Optional but useful
httpx

# Scheduler + testing
apscheduler
pytest
celery
redis
